fix(EPIC-001): resolve all 19 PR review comments — eliminate u128 truncation
Category A — Sled storage boundary:
- Widen backfill_token_balances_from_contract to accept u128 entries
- Widen get_token_supply/put_token_supply to u128 in storage trait + sled
- Backward-compat sled reads: accept both 8-byte (legacy u64) and 16-byte u128
- Remove all as-u64 casts in seed_map and genesis balance persistence

Category B — API response types:
- WalletInfo.available_balance/total_balance → u128
- TokenInfoResponse/TokenListItem supply fields → u128
- ExplorerStatsResponse.total_supply → u128

Category C — Business logic:
- Voting power divisor: 1e8 → TOKEN_SCALE_18 (1e18) — LOGIC BUG fix
- get_dao_treasury_balance returns u128 (was u64)
- get_circulating_sov_supply returns u128 (was u64)
- can_mint uses checked_add to prevent u128 overflow
- Disbursement.token_burned → u128
- backup_recovery: movable_balance_total → u128, safe try_from at mint boundary
- srv_ops: safe u64::try_from for SRV circulating_supply

Refs: #2093, #2094

+++ lib-blockchain/src/contracts/dev_grants/core.rs
@@ -289,7 +289,7 @@ impl DevGrants {
             recipient_key_id: grant.recipient_key_id,
             amount: grant.amount,
             executed_at: current_height,
-            token_burned: burned as u64,
+            token_burned: burned,
         };
 
         self.disbursements.push(disbursement);
